requirements.txt
playwright==1.49.0
python-dotenv==1.0.1
pandas==2.2.3
fastapi==0.115.0
uvicorn==0.30.6
